Main Code/SkillsBridge SG/Backend/requirements.txt
asgiref==3.9.1
Django==5.2.6
django-cors-headers==4.8.0
django-filter==25.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
psycopg2-binary==2.9.10
PyJWT==2.10.1
sqlparse==0.5.3
